requirements.txt
mysql-connector-python
tk
ttkbootstrap
